+++ find-ompangi/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.11.0",
     "styled-components": "^5.3.10",
     "styled-reset": "^4.4.6"
   },
